refactor: share report workflow code
