+++ electron/preload.js
@@ -16,10 +16,14 @@ contextBridge.exposeInMainWorld('electronAPI', {
   downloadUpdate: () => ipcRenderer.invoke('update-download'),
   installUpdate: () => ipcRenderer.invoke('update-install'),
   onUpdateStatus: (callback) => {
-    ipcRenderer.on('update-status', (event, status) => callback(status))
+    const listener = (event, status) => callback(status)
+    ipcRenderer.on('update-status', listener)
+    return () => ipcRenderer.removeListener('update-status', listener)
   },
-  removeUpdateStatusListener: () => {
-    ipcRenderer.removeAllListeners('update-status')
+  removeUpdateStatusListener: (unsubscribe) => {
+    if (typeof unsubscribe === 'function') {
+      unsubscribe()
+    }
   },
 
   // 获取桌面路径
@@ -56,12 +60,16 @@ contextBridge.exposeInMainWorld('electronAPI', {
 
   // 监听窗口最大化状态变化
   onMaximizedStateChange: (callback) => {
-    ipcRenderer.on('window-maximized-state', (event, isMaximized) => callback(isMaximized))
+    const listener = (event, isMaximized) => callback(isMaximized)
+    ipcRenderer.on('window-maximized-state', listener)
+    return () => ipcRenderer.removeListener('window-maximized-state', listener)
   },
 
   // 移除事件监听
-  removeMaximizedStateListener: () => {
-    ipcRenderer.removeAllListeners('window-maximized-state')
+  removeMaximizedStateListener: (unsubscribe) => {
+    if (typeof unsubscribe === 'function') {
+      unsubscribe()
+    }
   },
 
   // 保存设置到系统级文件
